refactor: changing the distribution by packages
Separate presentation, data and domain layer. Add repository interface
in domain.
Add packages, that I will use in future.
