+++ app/src/main/java/com/arekalov/yandexshmr/presentation/edit/EditScreen.kt
@@ -1,4 +1,4 @@
-package com.arekalov.yandexshmr.screens
+package com.arekalov.yandexshmr.presentation.edit
 
 import android.content.res.Configuration
 import androidx.compose.foundation.ScrollState
@@ -47,11 +47,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.arekalov.yandexshmr.R
-import com.arekalov.yandexshmr.ToDoItemsViewModel
-import com.arekalov.yandexshmr.models.Priority
-import com.arekalov.yandexshmr.models.ToDoItem
-import com.arekalov.yandexshmr.navigation.NEW_ITEM
-import com.arekalov.yandexshmr.ui.ToDoListTheme
+import com.arekalov.yandexshmr.domain.Priority
+import com.arekalov.yandexshmr.domain.ToDoItem
+import com.arekalov.yandexshmr.presentation.ToDoItemsViewModel
+import com.arekalov.yandexshmr.presentation.common.navigation.NEW_ITEM
+import com.arekalov.yandexshmr.presentation.theme.ToDoListTheme
 import java.time.Instant
 import java.time.LocalDate
 import java.time.LocalDateTime
