+++ app/src/main/java/com/arekalov/yandexshmr/presentation/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.arekalov.yandexshmr.ui
+package com.arekalov.yandexshmr.presentation.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
